Extract isAdmin strategy to authorization helper class
- The isAdmin method will be used in several authorization methods
